refactor(github): split issue/PR readers out of api.rs and move IssueComment to types.rs
Move list_issues/list_prs/read_issue/read_pr/fetch_issue_comments into a
focused issues.rs, keep pagination + merge_commit_batches in api.rs, and
relocate IssueComment to types.rs so the shared reader types live together.
Call sites in github.rs route through the new issues module; merge_commit_batches
gains unit tests for dedupe/ordering.
